Renamed and->all and or->any
