+++ pipeline-model-definition/src/test/resources/errors/whenNestedChildrenInvalid.groovy
@@ -35,8 +35,8 @@ pipeline {
         }
         stage("Two") {
             when {
-                and {
-                    and {
+                all {
+                    all {
 
                     }
                     not {
