+++ pipeline-model-definition/src/test/resources/json/whenOr.json
@@ -32,7 +32,7 @@
         }]
       }],
       "when": {"conditions": [      {
-        "name": "or",
+        "name": "any",
         "children":         [
           {
             "name": "branch",
